refactor(core): Move execution error creation to execution service (no-changelog)
