+++ packages/cli/src/GenericHelpers.ts
@@ -1,21 +1,11 @@
 import type express from 'express';
-import type {
-	ExecutionError,
-	INode,
-	IRunExecutionData,
-	Workflow,
-	WorkflowExecuteMode,
-} from 'n8n-workflow';
 import { validate } from 'class-validator';
-import { Container } from 'typedi';
 import config from '@/config';
-import type { ExecutionPayload, IWorkflowDb } from '@/Interfaces';
 import type { WorkflowEntity } from '@db/entities/WorkflowEntity';
 import type { CredentialsEntity } from '@db/entities/CredentialsEntity';
 import type { TagEntity } from '@db/entities/TagEntity';
 import type { User } from '@db/entities/User';
 import type { UserUpdatePayload } from '@/requests';
-import { ExecutionRepository } from '@db/repositories/execution.repository';
 import { BadRequestError } from './errors/response-errors/bad-request.error';
 
 /**
@@ -56,87 +46,6 @@ export async function validateEntity(
 	if (errorMessages) {
 		throw new BadRequestError(errorMessages);
 	}
-}
-
-/**
- * Create an error execution
- *
- * @param {INode} node
- * @param {IWorkflowDb} workflowData
- * @param {Workflow} workflow
- * @param {WorkflowExecuteMode} mode
- * @returns
- * @memberof ActiveWorkflowRunner
- */
-
-export async function createErrorExecution(
-	error: ExecutionError,
-	node: INode,
-	workflowData: IWorkflowDb,
-	workflow: Workflow,
-	mode: WorkflowExecuteMode,
-): Promise<void> {
-	const saveDataErrorExecutionDisabled = workflowData?.settings?.saveDataErrorExecution === 'none';
-
-	if (saveDataErrorExecutionDisabled) return;
-
-	const executionData: IRunExecutionData = {
-		startData: {
-			destinationNode: node.name,
-			runNodeFilter: [node.name],
-		},
-		executionData: {
-			contextData: {},
-			metadata: {},
-			nodeExecutionStack: [
-				{
-					node,
-					data: {
-						main: [
-							[
-								{
-									json: {},
-									pairedItem: {
-										item: 0,
-									},
-								},
-							],
-						],
-					},
-					source: null,
-				},
-			],
-			waitingExecution: {},
-			waitingExecutionSource: {},
-		},
-		resultData: {
-			runData: {
-				[node.name]: [
-					{
-						startTime: 0,
-						executionTime: 0,
-						error,
-						source: [],
-					},
-				],
-			},
-			error,
-			lastNodeExecuted: node.name,
-		},
-	};
-
-	const fullExecutionData: ExecutionPayload = {
-		data: executionData,
-		mode,
-		finished: false,
-		startedAt: new Date(),
-		workflowData,
-		workflowId: workflow.id,
-		stoppedAt: new Date(),
-		status: 'error',
-	};
-
-	await Container.get(ExecutionRepository).createNewExecution(fullExecutionData);
 }
 
 export const DEFAULT_EXECUTIONS_GET_ALL_LIMIT = 20;
