+++ packages/cli/src/executions/executions.service.ts
@@ -1,5 +1,13 @@
 import { validate as jsonSchemaValidate } from 'jsonschema';
-import type { IWorkflowBase, JsonObject, ExecutionStatus } from 'n8n-workflow';
+import type {
+	IWorkflowBase,
+	JsonObject,
+	ExecutionStatus,
+	ExecutionError,
+	INode,
+	IRunExecutionData,
+	WorkflowExecuteMode,
+} from 'n8n-workflow';
 import { ApplicationError, jsonParse, Workflow, WorkflowOperationError } from 'n8n-workflow';
 import type { FindOperator } from 'typeorm';
 import { In } from 'typeorm';
@@ -7,9 +15,11 @@ import { ActiveExecutions } from '@/ActiveExecutions';
 import config from '@/config';
 import type { User } from '@db/entities/User';
 import type {
+	ExecutionPayload,
 	IExecutionFlattedResponse,
 	IExecutionResponse,
 	IExecutionsListResponse,
+	IWorkflowDb,
 	IWorkflowExecutionDataProcess,
 } from '@/Interfaces';
 import { NodeTypes } from '@/NodeTypes';
@@ -361,5 +371,76 @@ export class ExecutionsService {
 				ids,
 			},
 		);
+	}
+
+	async createErrorExecution(
+		error: ExecutionError,
+		node: INode,
+		workflowData: IWorkflowDb,
+		workflow: Workflow,
+		mode: WorkflowExecuteMode,
+	): Promise<void> {
+		const saveDataErrorExecutionDisabled =
+			workflowData?.settings?.saveDataErrorExecution === 'none';
+
+		if (saveDataErrorExecutionDisabled) return;
+
+		const executionData: IRunExecutionData = {
+			startData: {
+				destinationNode: node.name,
+				runNodeFilter: [node.name],
+			},
+			executionData: {
+				contextData: {},
+				metadata: {},
+				nodeExecutionStack: [
+					{
+						node,
+						data: {
+							main: [
+								[
+									{
+										json: {},
+										pairedItem: {
+											item: 0,
+										},
+									},
+								],
+							],
+						},
+						source: null,
+					},
+				],
+				waitingExecution: {},
+				waitingExecutionSource: {},
+			},
+			resultData: {
+				runData: {
+					[node.name]: [
+						{
+							startTime: 0,
+							executionTime: 0,
+							error,
+							source: [],
+						},
+					],
+				},
+				error,
+				lastNodeExecuted: node.name,
+			},
+		};
+
+		const fullExecutionData: ExecutionPayload = {
+			data: executionData,
+			mode,
+			finished: false,
+			startedAt: new Date(),
+			workflowData,
+			workflowId: workflow.id,
+			stoppedAt: new Date(),
+			status: 'error',
+		};
+
+		await Container.get(ExecutionRepository).createNewExecution(fullExecutionData);
 	}
 }
